+++ assets/javascripts/kitten/components/meters/progress/stories.js
@@ -1,7 +1,12 @@
 import React from 'react'
-import { number, color } from '@storybook/addon-knobs'
+import { number, color, select } from '@storybook/addon-knobs'
 import { Progress as _Progress } from './index'
 import COLORS from '../../../constants/colors-config'
+
+const variantOptions = {
+  Andromeda: 'andromeda',
+  Orion: 'orion',
+}
 
 export default {
   title: 'Meters/Progress',
@@ -14,5 +19,6 @@ export const Progress = args => (
     value={number('Value', 50)}
     color={color('Color', COLORS.primary1)}
     rampProps={{ style: { height: 6 } }}
+    variant={select('Variant', variantOptions, 'andromeda')}
   />
 )